pyproject.toml
[project]
name = "first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Owais", email = "afshanaslam091@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai>=1.42.0",
    "openai-agents>=0.2.10",
    "python-dotenv>=1.1.1",
]

[project.scripts]
first_agent = "first_agent:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
